Split lib core responsibilties
